RepoBiteCoFinal/biteco_fixed/Base_Datos/requirements.txt
django
djangorestframework
psycopg2-binary
requests
gunicorn
django-redis
redis
PyJWT
